Refactor the Cellpose 3 classes.
Make an abstract detector, so that we can reuse if for Cellpose-SAM.
And move them into a dedicated package.
